+++ ElvUI_MerathilisUI/Core/Modules/Misc/Loot.lua
@@ -1,6 +1,6 @@
 local MER, F, E, L, V, P, G = unpack(select(2, ...))
 local module = MER:GetModule('MER_Misc')
-local LCG = LibStub('LibCustomGlow-1.0')
+local LCG = E.Libs.CustomGlow
 local M = E:GetModule('Misc')
 
 local _G = _G