+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Changes ###
 
-  * Fixed a C-Stack error
+  * Fixed an error with LibOpenRaid
+  * Removed LibCustomGlow, its back in ElvUI
+  * Updated ExpansionLandingPage Skin
+  * Temp workaround for EquipSets in Bags
+  * Fixed an error on Mail Frame